+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,4 @@
+{
+  "ExpandedNodes": [],
+  "PreviewInSolutionExplorer": false
+}
