+++ checker.csproj
@@ -1,6 +1,7 @@
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <Compile Include="checker.cs"/>
+    <Compile Include="Vitals/checker.cs"/>
+	<Compile Include="Vitals/VitalsChecker.cs"/>
   </ItemGroup>
   <Target Name="Build">
     <Csc Sources="@(Compile)"/>
